refactor: refactoring module dependency
#49
