+++ src/rotation/rotations.module.ts
@@ -1,40 +1,31 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { ScheduleModule } from '@nestjs/schedule';
-import { RotationEntity } from './entity/rotation.entity';
-import { RotationAttendeeEntity } from './entity/rotation-attendee.entity';
 import { RotationsService } from './rotations.service';
 import { RotationsController } from './rotations.controller';
-import { CustomRotationRepository } from './repository/rotations.repository';
-import { RotationHolidayEntity } from '../holiday/entity/holiday.entity';
-import { HolidayService } from '../holiday/holiday.service';
-import { HolidayRepository } from '../holiday/repository/holiday.repository';
+import { RotationRepository } from './repository/rotations.repository';
 import { HolidayModule } from '../holiday/holiday.module';
 import { UserService } from 'src/user/user.service';
 import { UserRepository } from 'src/user/repository/user.repository';
-import { UserModule } from 'src/user/user.module';
+import { RotationAttendeeRepository } from './repository/rotation-attendees.repository';
+import { RotationEntity } from './entity/rotation.entity';
+import { RotationAttendeeEntity } from './entity/rotation-attendee.entity';
 import { UserEntity } from 'src/user/entity/user.entity';
 
 @Module({
   imports: [
-    TypeOrmModule.forFeature([
-      RotationEntity,
-      RotationAttendeeEntity,
-      RotationHolidayEntity,
-      UserEntity,
-    ]),
+    TypeOrmModule.forFeature([RotationEntity, RotationAttendeeEntity, UserEntity]),
     ScheduleModule.forRoot(),
-    UserModule,
     HolidayModule,
   ],
   controllers: [RotationsController],
   providers: [
+    RotationsService,
+    RotationRepository,
+    RotationAttendeeRepository,
     UserService,
     UserRepository,
-    RotationsService,
-    CustomRotationRepository,
-    HolidayService,
-    HolidayRepository,
   ],
+  exports: [RotationsService, RotationRepository, RotationAttendeeRepository],
 })
 export class RotationsModule {}
