+++ src/rotation/utils/types.ts
@@ -1,0 +1,12 @@
+export interface DayObject {
+  day: number;
+  arr: number[];
+}
+
+export interface RotationAttendeeInfo {
+  userId: number;
+  year: number;
+  month: number;
+  attendLimit: number[];
+  attended: number;
+}
