+++ src/user/user.module.ts
@@ -4,11 +4,26 @@ import { UserController } from './user.controller';
 import { UserService } from './user.service';
 import { UserRepository } from './repository/user.repository';
 import { UserEntity } from './entity/user.entity';
+import { RotationsService } from '../rotation/rotations.service';
+import { RotationEntity } from 'src/rotation/entity/rotation.entity';
+import { RotationAttendeeEntity } from 'src/rotation/entity/rotation-attendee.entity';
+import { RotationRepository } from 'src/rotation/repository/rotations.repository';
+import { RotationAttendeeRepository } from 'src/rotation/repository/rotation-attendees.repository';
+import { HolidayModule } from 'src/holiday/holiday.module';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([UserEntity])],
+  imports: [
+    TypeOrmModule.forFeature([UserEntity, RotationEntity, RotationAttendeeEntity]),
+    HolidayModule,
+  ],
   controllers: [UserController],
-  providers: [UserService, UserRepository],
+  providers: [
+    UserService,
+    UserRepository,
+    RotationsService,
+    RotationRepository,
+    RotationAttendeeRepository,
+  ],
   exports: [UserService, UserRepository],
 })
 export class UserModule {}
